refactor(workbench): 拆分 workbenchDocumentSlice 守巨壳门岗
workbenchStore.ts 被 P2/P4 的状态扩展撑到 883 行，触发 check:filesize（R9/R12）。
按 generationCanvasStore 既有的 zustand slice 模式，抽出「创作文档 + 分镜方案」块
为 workbenchDocumentSlice.ts（172 行），workbenchStore.ts 回到 740 行。

纯结构拆分，行为不变：单测 2989 全绿。
