+++ src/workbench/workbenchStore.ts
@@ -32,18 +32,12 @@ import type { Vec2 } from './timeline/overlayTransform'
 import { createDefaultTimeline, normalizeTimeline } from './timeline/timelineMath'
 import { readPreviewSourceCollapsed, writePreviewSourceCollapsed } from './preview/previewSourcePanelPreference'
 import type { TimelineClip, TimelineState, TimelineTextStyle, TimelineTrackType } from './timeline/timelineTypes'
-import { createDefaultWorkbenchDocument, normalizeWorkbenchDocument, type CreationDocumentTools, type PreviewAspectRatio, type WorkbenchDocument } from './workbenchTypes'
+import { normalizeWorkbenchDocument, type CreationDocumentTools, type PreviewAspectRatio, type WorkbenchDocument } from './workbenchTypes'
 import type { WorkbenchAiMessage } from './ai/workbenchAiTypes'
-import type { StoryboardPlan } from './generationCanvas/agent/storyboardPlan'
 import type { ComposerAttachment } from './ai/composer/composerAttachmentTypes'
 import { createConversationBuckets } from './aiConversationBuckets'
 import { abandonCreationTurn } from './creation/creationTurnController'
-
-/** 一篇原稿的分镜方案条目（P4：storyboardPlans 的值）。 */
-export type StoryboardPlanEntry = {
-  plan: StoryboardPlan
-  committed: boolean
-}
+import { createWorkbenchDocumentSlice, type WorkbenchDocumentSlice } from './workbenchDocumentSlice'
 
 // 创作面板会话「会话域」per-project 桶(S1 治串台)。
 // 注:messages 已迁出本桶,改由 conversationThreads 模型按项目寻址(会话历史,2026-06-14);
@@ -69,8 +63,6 @@ export type TimelineSnapGuide = { frame: number; label: string }
 
 // 时间轴撤销栈封顶（防无限增长）。
 const TIMELINE_UNDO_LIMIT = 30
-// 初始文档（模块级单例）：保证 workbenchDocuments[0] 与 activeDocumentId 指向同一篇。
-const INITIAL_DOCUMENT = createDefaultWorkbenchDocument()
 // 离散编辑生效时把旧 timeline 压栈：仅当真的变了。供 set 内联调用。
 function pushTimelineUndo(stack: TimelineState[], previous: TimelineState): TimelineState[] {
   const next = [...stack, previous]
@@ -84,7 +76,7 @@ export type WorkspaceMode = (typeof WORKSPACE_MODES)[number]
 
 type GraphViewport = { zoom: number; offset: { x: number; y: number } }
 
-type WorkbenchState = {
+type WorkbenchState = WorkbenchDocumentSlice & {
   persistRevision: number
   workspaceMode: WorkspaceMode
   /** 生成/预览区右侧助手侧栏宽度（px，可拖宽）。 */
@@ -112,10 +104,6 @@ type WorkbenchState = {
   toggleSidebarCollapsed: () => void
   setSidebarCollapsed: (collapsed: boolean) => void
   rememberCategoryViewport: (categoryId: string, viewport: GraphViewport) => void
-  /** 原稿文档集合（有序，多文档侧栏真相源）。随项目持久化。 */
-  workbenchDocuments: WorkbenchDocument[]
-  /** 当前激活的原稿文档 id（切换文档 = 切编辑器内容）。随项目持久化。 */
-  activeDocumentId: string
   creationDocumentTools: CreationDocumentTools | null
   creationSelectionText: string; storyboardPlannerLauncher: ((displayPrompt?: string) => void) | null
   creationAiModeId: string
@@ -125,8 +113,6 @@ type WorkbenchState = {
   creationAiMessages: WorkbenchAiMessage[]
   creationAiAttachments: ComposerAttachment[]
   creationAiError: string
-  /** 每篇原稿的分镜方案（P4：按 documentId 索引，切原稿切方案，不再串稿）。随项目持久化。 */
-  storyboardPlans: Record<string, StoryboardPlanEntry>
   /**
    * 「请画布适应视图」一次性信号（nonce，仿 createCategoryNonce）。bump 一次 = 请生成画布
    * 平滑 fit 到全部节点一次。用于落画布等「批量加节点到已加载画布」的场景——useAutoFitOnLoad
@@ -166,18 +152,6 @@ type WorkbenchState = {
   setWorkspaceMode: (mode: unknown) => void
   setAssistantWidth: (width: number) => void
   setProjectSidebarWidth: (width: number) => void
-  /** 更新某篇文档（按 id 定位），并 bump 持久化。 */
-  setWorkbenchDocument: (document: WorkbenchDocument) => void
-  /** 新增一篇原稿（默认空白），返回新文档并设为激活。 */
-  addWorkbenchDocument: () => WorkbenchDocument
-  /** 删除一篇原稿（不可删到 0 篇，至少保留一篇空稿）。 */
-  deleteWorkbenchDocument: (id: string) => void
-  /** 改名一篇原稿（空名忽略）。 */
-  renameWorkbenchDocument: (id: string, title: string) => void
-  /** 切换激活文档（id 不存在则忽略）。 */
-  setActiveDocumentId: (id: string) => void
-  /** 恢复整套文档集合 + 激活 id（项目载入专用，不标脏）。 */
-  hydrateWorkbenchDocuments: (documents: WorkbenchDocument[], activeId: string | null) => void
   setCreationDocumentTools: (tools: CreationDocumentTools | null) => void
   setCreationSelectionText: (text: string) => void; setStoryboardPlannerLauncher: (launcher: ((displayPrompt?: string) => void) | null) => void
   setCreationAiModeId: (modeId: string) => void
@@ -186,16 +160,8 @@ type WorkbenchState = {
   setCreationAiMessages: (messages: WorkbenchAiMessage[] | ((messages: WorkbenchAiMessage[]) => WorkbenchAiMessage[])) => void
   setCreationAiAttachments: (attachments: ComposerAttachment[] | ((attachments: ComposerAttachment[]) => ComposerAttachment[])) => void
   setCreationAiError: (error: string) => void
-  /** 写入/改写分镜方案对象（planner 落库、编辑器逐字段编辑）：置草稿态。按 documentId 索引；缺省回退 activeDocumentId。 */
-  setStoryboardPlan: (plan: StoryboardPlan | null, documentId?: string) => void
-  /** 确认落画布后：方案保留、转「已落画布」（卡片留痕）。按 documentId 索引；缺省回退 activeDocumentId。 */
-  commitStoryboardPlan: (documentId?: string) => void
   /** 请生成画布平滑 fit 一次；可显式切到并绑定目标分类。 */
   requestCanvasFit: (categoryId?: string) => void
-  /** 丢弃方案：清空该文档的方案（卡片随之消失）。按 documentId 索引；缺省回退 activeDocumentId。 */
-  discardStoryboardPlan: (documentId?: string) => void
-  /** 项目载入专用：恢复整套方案映射，不标脏（区别于用户动作 setStoryboardPlan）。 */
-  hydrateStoryboardPlans: (entries: Record<string, StoryboardPlanEntry>) => void
   /** 切项目时交换对话桶(S1 治串台):存旧项目的对话,载入新项目的(没有则空)。 */
   swapCreationAiProject: (prevId: string | null, nextId: string | null) => void
   /** 一次性信号：打开示例/新项目时请求创作助手默认展开（让「拆镜头」CTA 一眼可见），消费后清掉。 */
@@ -260,7 +226,8 @@ export function isWorkspaceMode(value: unknown): value is WorkspaceMode {
   return typeof value === 'string' && WORKSPACE_MODES.includes(value as WorkspaceMode)
 }
 
-export const useWorkbenchStore = create<WorkbenchState>()(subscribeWithSelector((set, get) => ({
+export const useWorkbenchStore = create<WorkbenchState>()(subscribeWithSelector((set, get, store) => ({
+  ...createWorkbenchDocumentSlice(set, get, store),
   persistRevision: 0,
   workspaceMode: 'generation',
   assistantWidth: 340,
@@ -330,8 +297,6 @@ export const useWorkbenchStore = create<WorkbenchState>()(subscribeWithSelector(
       },
     }))
   },
-  workbenchDocuments: [INITIAL_DOCUMENT],
-  activeDocumentId: INITIAL_DOCUMENT.id,
   creationDocumentTools: null,
   creationSelectionText: '', storyboardPlannerLauncher: null,
   creationAiModeId: 'general',
@@ -340,7 +305,6 @@ export const useWorkbenchStore = create<WorkbenchState>()(subscribeWithSelector(
   creationAiMessages: [],
   creationAiAttachments: [],
   creationAiError: '',
-  storyboardPlans: {},
   canvasFitNonce: 0,
   canvasFitCategoryId: null,
   creationAssistantAutoOpen: false,
@@ -367,68 +331,6 @@ export const useWorkbenchStore = create<WorkbenchState>()(subscribeWithSelector(
   },
   setAssistantWidth: (width) => set({ assistantWidth: Math.max(300, Math.min(600, Math.round(width))) }),
   setProjectSidebarWidth: (width) => set({ projectSidebarWidth: Math.max(240, Math.min(720, Math.round(width))) }),
-  setWorkbenchDocument: (workbenchDocument) => {
-    const normalized = normalizeWorkbenchDocument(workbenchDocument)
-    set((state) => {
-      const exists = state.workbenchDocuments.some((d) => d.id === normalized.id)
-      return {
-        workbenchDocuments: exists
-          ? state.workbenchDocuments.map((d) => (d.id === normalized.id ? normalized : d))
-          : [...state.workbenchDocuments, normalized],
-        activeDocumentId: exists ? state.activeDocumentId : normalized.id,
-        persistRevision: state.persistRevision + 1,
-      }
-    })
-  },
-  addWorkbenchDocument: () => {
-    const doc = createDefaultWorkbenchDocument()
-    set((state) => ({
-      workbenchDocuments: [...state.workbenchDocuments, doc],
-      activeDocumentId: doc.id,
-      persistRevision: state.persistRevision + 1,
-    }))
-    return doc
-  },
-  deleteWorkbenchDocument: (id) => {
-    if (typeof id !== 'string' || !id.trim()) return
-    set((state) => {
-      if (state.workbenchDocuments.length <= 1) return state // 至少保留一篇
-      const target = state.workbenchDocuments.find((d) => d.id === id)
-      if (!target) return state
-      const next = state.workbenchDocuments.filter((d) => d.id !== id)
-      const nextActive = state.activeDocumentId === id ? next[0].id : state.activeDocumentId
-      return {
-        workbenchDocuments: next,
-        activeDocumentId: nextActive,
-        persistRevision: state.persistRevision + 1,
-      }
-    })
-  },
-  renameWorkbenchDocument: (id, title) => {
-    const trimmed = (title || '').trim()
-    if (!trimmed) return
-    set((state) => {
-      if (!state.workbenchDocuments.some((d) => d.id === id)) return state
-      return {
-        workbenchDocuments: state.workbenchDocuments.map((d) => (d.id === id ? { ...d, title: trimmed, updatedAt: Date.now() } : d)),
-        persistRevision: state.persistRevision + 1,
-      }
-    })
-  },
-  setActiveDocumentId: (id) => {
-    if (typeof id !== 'string' || !id.trim()) return
-    set((state) => {
-      if (!state.workbenchDocuments.some((d) => d.id === id)) return state
-      if (state.activeDocumentId === id) return state
-      return { activeDocumentId: id }
-    })
-  },
-  hydrateWorkbenchDocuments: (documents, activeId) => {
-    const normalized = documents.map(normalizeWorkbenchDocument)
-    const safe = normalized.length ? normalized : [createDefaultWorkbenchDocument()]
-    const active = safe.some((d) => d.id === activeId) ? (activeId as string) : safe[0].id
-    set({ workbenchDocuments: safe, activeDocumentId: active })
-  },
   setCreationDocumentTools: (creationDocumentTools) => {
     set({ creationDocumentTools })
   },
@@ -458,57 +360,12 @@ export const useWorkbenchStore = create<WorkbenchState>()(subscribeWithSelector(
   setCreationAiError: (creationAiError) => {
     set({ creationAiError })
   },
-  setStoryboardPlan: (storyboardPlan, documentId) => {
-    // P0-6:方案是 per-project 持久化产物 → bump persistRevision 触发防抖落盘(否则用户手改的方案不保存)。
-    // 写/改方案一律置草稿态(被编辑即与画布上旧节点不一致)。P4:按 documentId 索引（缺省回退激活文档）。
-    const target = typeof documentId === 'string' && documentId.trim() ? documentId.trim() : get().activeDocumentId
-    if (!target) return
-    set((state) => {
-      if (storyboardPlan === null) {
-        if (!state.storyboardPlans[target]) return state
-        const next = { ...state.storyboardPlans }
-        delete next[target]
-        return { storyboardPlans: next, persistRevision: state.persistRevision + 1 }
-      }
-      return {
-        storyboardPlans: { ...state.storyboardPlans, [target]: { plan: storyboardPlan, committed: false } },
-        persistRevision: state.persistRevision + 1,
-      }
-    })
-  },
-  commitStoryboardPlan: (documentId) => {
-    // 确认落画布:方案保留(卡片留痕)、转已落画布。bump 落盘 committed 状态。
-    const target = typeof documentId === 'string' && documentId.trim() ? documentId.trim() : get().activeDocumentId
-    if (!target) return
-    set((state) => {
-      const entry = state.storyboardPlans[target]
-      if (!entry || entry.committed) return state
-      return {
-        storyboardPlans: { ...state.storyboardPlans, [target]: { ...entry, committed: true } },
-        persistRevision: state.persistRevision + 1,
-      }
-    })
-  },
-  discardStoryboardPlan: (documentId) => {
-    const target = typeof documentId === 'string' && documentId.trim() ? documentId.trim() : get().activeDocumentId
-    if (!target) return
-    set((state) => {
-      if (!state.storyboardPlans[target]) return state
-      const next = { ...state.storyboardPlans }
-      delete next[target]
-      return { storyboardPlans: next, persistRevision: state.persistRevision + 1 }
-    })
-  },
   requestCanvasFit: (categoryId) => {
     // 一次性信号：目标分类与 nonce 原子更新。显式目标立即切过去，延迟消费时若用户又手动切走则跳过。
     set((state) => {
       const target = typeof categoryId === 'string' && categoryId.trim() ? categoryId.trim() : state.activeCategoryId
       return { activeCategoryId: target, canvasFitCategoryId: target, canvasFitNonce: state.canvasFitNonce + 1 }
     })
-  },
-  hydrateStoryboardPlans: (entries) => {
-    // 载入态:一次性设整套映射、不 bump persistRevision(restore 非用户编辑,别标脏触发回存)。
-    set({ storyboardPlans: entries })
   },
   setCreationAssistantAutoOpen: (creationAssistantAutoOpen) => {
     set({ creationAssistantAutoOpen })
